Slash command work (#3937)
