+++ src/commands/Minion/daily.ts
@@ -1,6 +1,6 @@
 import { Time } from 'e';
 import * as fs from 'fs';
-import { CommandStore, KlasaMessage } from 'klasa';
+import { CommandStore, KlasaMessage, KlasaUser } from 'klasa';
 
 import { COINS_ID, Emoji, SupportServer } from '../../lib/constants';
 import pets from '../../lib/data/pets';
@@ -32,6 +32,19 @@ const options = {
 	errors: ['time']
 };
 
+export function isUsersDailyReady(user: KlasaUser): { isReady: true } | { isReady: false; durationUntilReady: number } {
+	const currentDate = new Date().getTime();
+	const lastVoteDate = user.settings.get(UserSettings.LastDailyTimestamp);
+	const difference = currentDate - lastVoteDate;
+
+	if (difference < Time.Hour * 12) {
+		const duration = Date.now() - (lastVoteDate + Time.Hour * 12);
+		return { isReady: false, durationUntilReady: duration };
+	}
+
+	return { isReady: true };
+}
+
 export default class DailyCommand extends BotCommand {
 	public constructor(store: CommandStore, file: string[], directory: string) {
 		super(store, file, directory, {
@@ -46,19 +59,16 @@ export default class DailyCommand extends BotCommand {
 
 	async run(msg: KlasaMessage) {
 		if (msg.channel.id === '342983479501389826') return;
-		await msg.author.settings.sync();
-		const currentDate = new Date().getTime();
-		const lastVoteDate = msg.author.settings.get(UserSettings.LastDailyTimestamp);
-		const difference = currentDate - lastVoteDate;
-
-		// If they have already claimed a daily in the past 12h
-		if (difference < Time.Hour * 12) {
-			const duration = formatDuration(Date.now() - (lastVoteDate + Time.Hour * 12));
-
-			return msg.channel.send(`**${Emoji.Diango} Diango says...** You can claim your next daily in ${duration}.`);
+		const check = isUsersDailyReady(msg.author);
+		if (!check.isReady) {
+			return msg.channel.send(
+				`**${Emoji.Diango} Diango says...** You can claim your next daily in ${formatDuration(
+					check.durationUntilReady
+				)}.`
+			);
 		}
 
-		await msg.author.settings.update(UserSettings.LastDailyTimestamp, currentDate);
+		await msg.author.settings.update(UserSettings.LastDailyTimestamp, new Date().getTime());
 
 		const trivia = triviaQuestions[Math.floor(Math.random() * triviaQuestions.length)];
