+++ src/mahoji/commands/minion.ts
@@ -9,11 +9,13 @@ import {
 	claimAchievementDiaryCommand
 } from '../lib/abstracted_commands/achievementDiaryCommand';
 import { bankBgCommand } from '../lib/abstracted_commands/bankBgCommand';
+import { cancelTaskCommand } from '../lib/abstracted_commands/cancelTaskCommand';
 import { crackerCommand } from '../lib/abstracted_commands/crackerCommand';
 import { Lampables, lampCommand } from '../lib/abstracted_commands/lampCommand';
-import { skillOption } from '../lib/mahojiCommandOptions';
+import { allUsableItems, useCommand } from '../lib/abstracted_commands/useCommand';
+import { ownedItemOption, skillOption } from '../lib/mahojiCommandOptions';
 import { OSBMahojiCommand } from '../lib/util';
-import { MahojiUserOption } from '../mahojiSettings';
+import { MahojiUserOption, mahojiUsersSettingsFetch } from '../mahojiSettings';
 
 export const minionCommand: OSBMahojiCommand = {
 	name: 'minion',
@@ -114,6 +116,23 @@ export const minionCommand: OSBMahojiCommand = {
 					max_value: 100_000
 				}
 			]
+		},
+		{
+			type: ApplicationCommandOptionType.Subcommand,
+			name: 'cancel',
+			description: 'Cancel your current trip.'
+		},
+		{
+			type: ApplicationCommandOptionType.Subcommand,
+			name: 'use',
+			description: 'Allows you to use items.',
+			options: [
+				{
+					...ownedItemOption(i => allUsableItems.has(i.id)),
+					required: true,
+					name: 'item'
+				}
+			]
 		}
 	],
 	run: async ({
@@ -126,8 +145,11 @@ export const minionCommand: OSBMahojiCommand = {
 		bankbg?: { name?: string };
 		cracker?: { user: MahojiUserOption };
 		lamp?: { item: string; quantity?: number; skill: string };
+		cancel?: {};
+		use?: { item: string };
 	}>) => {
 		const user = await globalClient.fetchUser(userID.toString());
+		const mahojiUser = await mahojiUsersSettingsFetch(user.id);
 
 		if (options.stats) {
 			return { embeds: [await minionStatsEmbed(user)] };
@@ -151,6 +173,10 @@ export const minionCommand: OSBMahojiCommand = {
 		if (options.lamp) {
 			return lampCommand(user, options.lamp.item, options.lamp.skill, options.lamp.quantity);
 		}
+
+		if (options.cancel) return cancelTaskCommand(mahojiUser, interaction);
+
+		if (options.use) return useCommand(mahojiUser, user, options.use.item);
 
 		return 'Unknown command';
 	}
